WIP: Camera refactor
